removed CHANNEL env in header, removed redundancy by getting the ENV during the member call, added (int)
Couldn't determine why CHANNEL kept throwing an error of NoneType, when in the .env file it is present. Forgot that getenv stores the files as strings and the channel is looking for an integer. In get_channel, I am converting to an integer the channel value found in the .env file.